Basic motion, heater and fan control
Adds the required pin assigments to enable the following on the D3 Pro:

- X1, X2, Y, Z1 and Z2 endstops
- X1, X2, Y, Z1 and Z2 stepper control
- Heater0, Heater1 and Bed heaters
- Heater and Part Cooling fans

On the D3 Pro, the hotend and part cooling fans for both printheads
share the same channels. One of the reasons for wanting to migrate my
D3 Pro to an open firmware was to configure independent control of
these fans. Since this is not possible, I probably won't do any more
work on this. I have a Duet board kicking around that I'll put to use
with this printer instead.